Change folder name to avoir space problem
