--- Alazar ATS9360/ALAZARAPI.py
@@ -1,38 +1,0 @@
-class alazarapi():
-	# Generated by h2py from C:\AlazarTech\ATS-SDK\6.1.0\Samples\Include\AlazarApi.h
-	FPGA_GETFIRST = (-1)
-	FPGA_GETNEXT = (-2)
-	FPGA_GETLAST = (-4)
-	ADMA_CLOCKSOURCE = 0x00000001
-	ADMA_CLOCKEDGE = 0x00000002
-	ADMA_SAMPLERATE = 0x00000003
-	ADMA_INPUTRANGE = 0x00000004
-	ADMA_INPUTCOUPLING = 0x00000005
-	ADMA_IMPUTIMPEDENCE = 0x00000006
-	ADMA_EXTTRIGGERED = 0x00000007
-	ADMA_CHA_TRIGGERED = 0x00000008
-	ADMA_CHB_TRIGGERED = 0x00000009
-	ADMA_TIMEOUT = 0x0000000A
-	ADMA_THISCHANTRIGGERED = 0x0000000B
-	ADMA_SERIALNUMBER = 0x0000000C
-	ADMA_SYSTEMNUMBER = 0x0000000D
-	ADMA_BOARDNUMBER = 0x0000000E
-	ADMA_WHICHCHANNEL = 0x0000000F
-	ADMA_SAMPLERESOLUTION = 0x00000010
-	ADMA_DATAFORMAT = 0x00000011
-	CSO_DISABLE = 0
-	CSO_ENABLE_DUMMY_CLOCK = 1
-	CSO_TRIGGER_LOW_DUMMY_CLOCK = 2
-	ADMA_EXTERNAL_STARTCAPTURE = 0x00000001
-	ADMA_ENABLE_RECORD_HEADERS = 0x00000008
-	ADMA_SINGLE_DMA_CHANNEL = 0x00000010
-	ADMA_ALLOC_BUFFERS = 0x00000020
-	ADMA_TRADITIONAL_MODE = 0x00000000
-	ADMA_CONTINUOUS_MODE = 0x00000100
-	ADMA_NPT = 0x00000200
-	ADMA_TRIGGERED_STREAMING = 0x00000400
-	ADMA_FIFO_ONLY_STREAMING = 0x00000800
-	ADMA_INTERLEAVE_SAMPLES = 0x00001000
-	ADMA_GET_PROCESSED_DATA = 0x00002000
-	CRA_MODE_DISABLE = 0
-	CRA_MODE_ENABLE_FPGA_AVE = 1
