--- Alazar ATS9360/ALAZARCMD.py
@@ -1,185 +1,0 @@
-class alazarcmd():
-
-	# Generated by h2py from C:\AlazarTech\ATS-SDK\6.1.0\Samples\Include\AlazarCmd.h
-	DATA_FORMAT_UNSIGNED = 0
-	DATA_FORMAT_SIGNED = 1
-	ECC_DISABLE = 0
-	ECC_ENABLE = 1
-	AUX_INPUT_LOW = 0
-	AUX_INPUT_HIGH = 1
-	EXT_TRIG_50_OHMS = 0
-	EXT_TRIG_300_OHMS = 1
-	PACK_DEFAULT = 0
-	PACK_8_BITS_PER_SAMPLE = 1
-	MASTER_CLOCK = 0x000000011
-	INTERNAL_CLOCK_SET_VCO = 0x000000012
-	CHANNEL_ALL = 0x00000000
-	CHANNEL_A = 0x00000001
-	CHANNEL_B = 0x00000002
-	CHANNEL_C = 0x00000004
-	CHANNEL_D = 0x00000008
-	CHANNEL_E = 0x00000010
-	CHANNEL_F = 0x00000020
-	CHANNEL_G = 0x00000040
-	CHANNEL_H = 0x00000080
-	SETDAC_INVALID_SETGET = 660
-	SETDAC_INVALID_CHANNEL = 661
-	SETDAC_INVALID_DACNAME = 662
-	SETDAC_INVALID_COUPLING = 663
-	SETDAC_INVALID_RANGE = 664
-	SETDAC_INVALID_IMPEDANCE = 665
-	SETDAC_BAD_GET_PTR = 667
-	SETDAC_INVALID_BOARDTYPE = 668
-	CSO_DUMMY_CLOCK_DISABLE = 0
-	CSO_DUMMY_CLOCK_TIMER = 1
-	CSO_DUMMY_CLOCK_EXT_TRIGGER = 2
-	CSO_DUMMY_CLOCK_TIMER_ON_TIMER_OFF = 3
-	STOS_OPTION_DEFER_START_CAPTURE = 1
-	SSM_DISABLE = 0
-	SSM_ENABLE = 1
-	CPF_REG_SIGNATURE = 0
-	CPF_REG_REVISION = 1
-	CPF_REG_VERSION = 2
-	CPF_REG_STATUS = 3
-	CPF_OPTION_DMA_DOWNLOAD = 1
-	CPF_DEVICE_UNKNOWN = 0
-	CPF_DEVICE_EP3SL50 = 1
-	CPF_DEVICE_EP3SE260 = 2
-	
-	#Sample Rate values
-	
-	SAMPLE_RATE_1KSPS = 0X00000001   
-	SAMPLE_RATE_2KSPS	=	0X00000002
-	SAMPLE_RATE_5KSPS	=	0X00000004
-	SAMPLE_RATE_10KSPS	=	0X00000008
-	SAMPLE_RATE_20KSPS	=	0X0000000A
-	SAMPLE_RATE_50KSPS	=	0X0000000C
-	SAMPLE_RATE_100KSPS	=	0X0000000E
-	SAMPLE_RATE_200KSPS	=	0X00000010
-	SAMPLE_RATE_500KSPS	=	0X00000012
-	SAMPLE_RATE_1MSPS	=	0X00000014
-	SAMPLE_RATE_2MSPS	=	0X00000018
-	SAMPLE_RATE_5MSPS	=	0X0000001A
-	SAMPLE_RATE_10MSPS	=	0X0000001C
-	SAMPLE_RATE_20MSPS	=	0X0000001E
-	SAMPLE_RATE_25MSPS	=	0X00000021
-	SAMPLE_RATE_50MSPS	=	0X00000022
-	SAMPLE_RATE_100MSPS	=	0X00000024
-	SAMPLE_RATE_125MSPS	=	0x00000025
-	SAMPLE_RATE_160MSPS	=	0x00000026
-	SAMPLE_RATE_180MSPS	=	0x00000027
-	SAMPLE_RATE_200MSPS	=	0X00000028
-	SAMPLE_RATE_250MSPS	=	0X0000002B
-	SAMPLE_RATE_300MSPS	=	300000000
-	SAMPLE_RATE_400MSPS	=	400000000
-	SAMPLE_RATE_500MSPS	=	500000000
-	SAMPLE_RATE_800MSPS	=	800000000
-	SAMPLE_RATE_1000MSPS=	1000000000
-	SAMPLE_RATE_1GSPS	=	SAMPLE_RATE_1000MSPS
-	SAMPLE_RATE_1200MSPS=	1200000000
-	SAMPLE_RATE_1500MSPS=	1500000000
-	SAMPLE_RATE_1600MSPS=	1600000000
-	SAMPLE_RATE_1800MSPS=   1800000000
-	SAMPLE_RATE_2000MSPS=	2000000000
-	SAMPLE_RATE_2GSPS	=	SAMPLE_RATE_2000MSPS
-	
-	#Clock Source
-	
-	INTERNAL_CLOCK		=	0x00000001   
-	EXTERNAL_CLOCK		=	0x00000002
-	FAST_EXTERNAL_CLOCK	=	0x00000002
-	MEDIMUM_EXTERNAL_CLOCK =	0x00000003 #TYPO
-	MEDIUM_EXTERNAL_CLOCK =	0x00000003
-	SLOW_EXTERNAL_CLOCK	 =	0x00000004
-	EXTERNAL_CLOCK_AC	=	0x00000005
-	EXTERNAL_CLOCK_DC	=	0x00000006
-	EXTERNAL_CLOCK_10MHz_REF = 0x00000007
-	INTERNAL_CLOCK_DIV_5 =	0x000000010
-	MASTER_CLOCK		=	0x000000011
-	INTERNAL_CLOCK_SET_VCO =	0x000000012
-	
-	#Clock Edge
-	
-	CLOCK_EDGE_RISING	=	0x00000000  	
-	CLOCK_EDGE_FALLING	=	0x00000001
-	
-
-	#Impedance Values
-	
-	IMPEDANCE_1M_OHM	=	0x00000001
-	IMPEDANCE_50_OHM	=	0x00000002
-	IMPEDANCE_75_OHM	=	0x00000004
-	IMPEDANCE_300_OHM	=	0x00000008
-	
-
-	#Input Ranges
-	
-	INPUT_RANGE_PM_20_MV =	0x00000001
-	INPUT_RANGE_PM_40_MV =	0x00000002
-	INPUT_RANGE_PM_50_MV =	0x00000003
-	INPUT_RANGE_PM_80_MV =	0x00000004
-	INPUT_RANGE_PM_100_MV =	0x00000005
-	INPUT_RANGE_PM_200_MV =	0x00000006
-	INPUT_RANGE_PM_400_MV =	0x00000007
-	INPUT_RANGE_PM_500_MV =	0x00000008
-	INPUT_RANGE_PM_800_MV =	0x00000009
-	INPUT_RANGE_PM_1_V	=	0x0000000A
-	INPUT_RANGE_PM_2_V	=	0x0000000B
-	INPUT_RANGE_PM_4_V	=	0x0000000C
-	INPUT_RANGE_PM_5_V	=	0x0000000D
-	INPUT_RANGE_PM_8_V	=	0x0000000E
-	INPUT_RANGE_PM_10_V	=	0x0000000F
-	INPUT_RANGE_PM_20_V	=	0x00000010
-	INPUT_RANGE_PM_40_V	=	0x00000011
-	INPUT_RANGE_PM_16_V	=	0x00000012
-	INPUT_RANGE_HIFI	=	0x00000020
-	INPUT_RANGE_PM_1_V_25 =	0x00000021
-	INPUT_RANGE_PM_2_V_5 =	0x00000025
-	INPUT_RANGE_PM_125_MV =	0x00000028
-	INPUT_RANGE_PM_250_MV =	0x00000030
-	
-
-	#Coupling Values
-	
-	AC_COUPLING		=		0x00000001
-	DC_COUPLING		=		0x00000002
-
-	#Trigger Engines
-	
-	TRIG_ENGINE_J	=		0x00000000
-	TRIG_ENGINE_K	=		0x00000001
-
-
-	#Trigger Engine Operation
-	
-	TRIG_ENGINE_OP_J	=		0x00000000
-	TRIG_ENGINE_OP_K		=	0x00000001
-	TRIG_ENGINE_OP_J_OR_K	=	0x00000002
-	TRIG_ENGINE_OP_J_AND_K	=	0x00000003
-	TRIG_ENGINE_OP_J_XOR_K	=	0x00000004
-	TRIG_ENGINE_OP_J_AND_NOT_K =	0x00000005
-	TRIG_ENGINE_OP_NOT_J_AND_K =	0x00000006
-
-	#Trigger Engine Sources
-	
-	TRIG_CHAN_A		=		0x00000000
-	TRIG_CHAN_B		=		0x00000001
-	TRIG_EXTERNAL	=		0x00000002
-	TRIG_DISABLE	=		0x00000003
-	TRIG_CHAN_C		=	0x00000004
-	TRIG_CHAN_D		=		0x00000005
-
-	#Trigger Slope
-	
-	TRIGGER_SLOPE_POSITIVE =	0x00000001
-	TRIGGER_SLOPE_NEGATIVE =	0x00000002
-
-	#External Trigger Attenuator Relay
-	
-	ETR_DIV5		=		0x00000000
-	ETR_X1		=			0x00000001
-	ETR_5V			=		0x00000000
-	ETR_1V		=			0x00000001
-	ETR_TTL		=			0x00000002
-	ETR_2V5		=			0x00000003
-
