+++ Alazar_ATS9360/ALAZARERROR.py
@@ -1,0 +1,4 @@
+class alazarerror():
+
+	# Generated by h2py from C:\AlazarTech\ATS-SDK\6.1.0\Samples\Include\AlazarError.h
+	API_RETURN_CODE_STARTS = 0x200
